+++ component/pl_mutex.cpp
@@ -27,10 +27,9 @@ esp_err_t Mutex::Lock(TickType_t timeout) {
   ESP_RETURN_ON_FALSE(!xPortInIsrContext(), ESP_ERR_INVALID_STATE, TAG, "calling mutex lock from ISR");
   if (xSemaphoreTakeRecursive(mutex, timeout))
     return ESP_OK;
-  if (timeout == 0)
-    return ESP_ERR_TIMEOUT;
-  ESP_RETURN_ON_ERROR(ESP_ERR_TIMEOUT, TAG, "timeout");
-  return ESP_OK;
+  if (timeout != 0)
+    ESP_LOGE(TAG, "timeout");
+  return ESP_ERR_TIMEOUT;
 }
 
 //==============================================================================