+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Simplified mutex and stream server lock timeout handling.
+
 ### Fixed
 - Event handler memory leak.
 - Missing lock in stream read buffer flushing.